refactor: split builderRenderer into managers
