+++ BlogposterCMS/public/assets/plainspace/builder/managers/widgetManager.js
@@ -1,0 +1,3 @@
+export { attachEditButton, attachRemoveButton, attachLockOnClick } from '../renderer/widgetActions.js';
+export { attachOptionsMenu } from '../widgets/widgetMenu.js';
+export { renderWidget } from '../widgets/widgetRenderer.js';
